fix: update で Int を返すよう修正

+++ app/src/main/java/app/doggy/newmybrary/data/db/dao/BookDao.kt
@@ -14,7 +14,7 @@ interface BookDao {
   fun insert(book: BookEntity): Long
 
   @Update
-  fun update(book: BookEntity): Long
+  fun update(book: BookEntity): Int
 
   @Transaction
   @Query("SELECT * FROM books")
